Allow to configure the number of persons to book for
This change introduces a new configuration variable, TOTAL_PERSONS,
which sets the number of people who need to book an appointment
together. Right now the IND allows a maximum of 6 people, with each
person's appointment duration set to 10 minutes. Multiple people booking
together need consecutive slots to be available.

Configuring the number of persons makes the bot much less noisy for
groups of people booking together because most of the slots that free up
are long enough for a single person.